aider: Add whisper-node to package.json dependencies.
# Aider chat conversation:

USER: não precisa add no `package.json`?

ASSISTANT: Certainly! To use the `whisper-node` package, it should be added to the `package.json` file. Here is the diff to add `whisper-node` to the dependencies:

```diff
--- package.json
+++ package.json
@@ ... @@
     "dependencies": {
       "@open-wa/wa-automate": "^4.67.0",
       "axios": "^1.6.7",
       "better-sqlite3": "^9.3.0",
       "openai": "^4.8.0",
       "form-data": "^4.0.0",
+      "whisper-node": "^1.0.0"
     }
   }
```

This change adds `whisper-node` to the list of dependencies with a version that is assumed to be compatible. You will need to run `npm install` after applying this patch to actually install the `whisper-node` package.